Social media marketing/backend/requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
tweepy
openai
python-dotenv
redis
cffi
clerk-backend-api
apscheduler
